agent/requirements.txt
tweepy
schedule
requests
python-dotenv
datetime
near-api
re